python/requirements.txt
Requests==2.23.0
python-dotenv==0.13.0